refactor: move `EVMFilterExec` to own file
